Split split-init into its own file
---
